Allow setting configuration, convert to context-based
This was a similar mistake to what I did with the input event provider, that had a separate set of events for each use of the hook. We should only watch history for changes once, so it doesn't have to do a ton of extra work when modifying the configuration.